Data-Science-Tutor/requirements.txt
streamlit==1.30.0
langchain==0.1.1
langchain-google-genai==0.0.6
google-generativeai==0.3.1
python-dotenv==1.0.0
streamlit-chat==0.1.1
google-api-python-client
matplotlib
pandas
graphviz